% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineup.R
\name{buildTaggle}
\alias{buildTaggle}
\title{factory for LineUp HTMLWidget based on a LineUpBuilder}
\usage{
buildTaggle(
  x,
  width = "100\%",
  height = NULL,
  elementId = NULL,
  dependencies = .crosstalkLineUpLibs()
)
}
\arguments{
\item{x}{LineUpBuilder object}

\item{width}{width of the element}

\item{height}{height of the element}

\item{elementId}{unique element id}

\item{dependencies}{include crosstalk dependencies}
}
\value{
taggle html widget
}
\description{
factory for LineUp HTMLWidget based on a LineUpBuilder
}
\examples{
\dontrun{
lineupBuilder(iris) |> buildTaggle()
}
}
